requirements.txt
numpy>=1.24
scipy>=1.11
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
Pillow>=10.0
tqdm>=4.65
jupyter>=1.0
ipykernel>=6.25
